Move renderImage() to api.ts
