+++ src/api.ts
@@ -1,6 +1,8 @@
 
 import * as zarr from "zarrita";
 
+import { Axis, Omero } from "./types/ome";
+import { getRgba, convertRbgDataToDataUrl } from "./render";
 import { NgffImage } from "./image";
 
 
@@ -29,4 +31,29 @@ export async function renderThumbnail(
   }
   let src = await ngffImg.render(targetSize, autoBoost);
   return src;
+}
+
+
+export async function renderImage(
+  arr: zarr.Array<any>,
+  axes: Axis[],
+  omero: Omero | null | undefined,
+  sliceIndices: { [k: string]: number | [number, number] | undefined } = {},
+  autoBoost: boolean = false,
+  originalShape?: number[]
+): Promise<string> {
+  // Main rendering function...
+  // We have the zarr Array already in hand, axes for dimensions
+  // and omero for rendering settings
+  // if autoBoost is true, check histogram and boost contrast if needed
+  let { data, width, height } = await getRgba(
+    arr,
+    axes,
+    omero,
+    sliceIndices,
+    originalShape,
+    autoBoost
+  );
+
+  return convertRbgDataToDataUrl(data, width, height);
 }
