+++ src/render.ts
@@ -11,31 +11,8 @@ import {
   MAX_CHANNELS,
 } from "./utils";
 
-export async function renderImage(
-  arr: zarr.Array<any>,
-  axes: Axis[],
-  omero: Omero | null | undefined,
-  sliceIndices: { [k: string]: number | [number, number] | undefined } = {},
-  autoBoost: boolean = false,
-  originalShape?: number[]
-): Promise<string> {
-  // Main rendering function...
-  // We have the zarr Array already in hand, axes for dimensions
-  // and omero for rendering settings
-  // if autoBoost is true, check histogram and boost contrast if needed
-  let { data, width, height } = await getRgba(
-    arr,
-    axes,
-    omero,
-    sliceIndices,
-    originalShape,
-    autoBoost
-  );
 
-  return convertRbgDataToDataUrl(data, width, height);
-}
-
-async function getRgba(
+export async function getRgba(
   arr: zarr.Array<any, zarr.Readable>,
   axes: Axis[],
   omero: Omero | null | undefined,
